Moved mocks into separate packages
