+++ test/controllers/ObligationsControllerSpec.scala
@@ -20,7 +20,7 @@ import assets.Messages.{ISE => errorMessages, Obligations => messages}
 import assets.TestConstants._
 import auth.MockAuthenticationPredicate
 import config.FrontendAppConfig
-import mocks.MockObligationsService
+import mocks.services.MockObligationsService
 import models._
 import org.jsoup.Jsoup
 import play.api.http.Status
