+++ test/mocks/services/MockEstimatedLiabilityService.scala
@@ -14,21 +14,20 @@
  * limitations under the License.
  */
 
-package mocks
+package mocks.services
 
-import models.{LastTaxCalculationResponseModel, ObligationsResponseModel}
+import models.LastTaxCalculationResponseModel
 import org.mockito.ArgumentMatchers
 import org.mockito.Mockito._
 import org.scalatest.BeforeAndAfterEach
 import org.scalatest.mockito.MockitoSugar
-import services.{EstimatedTaxLiabilityService, ObligationsService}
+import services.EstimatedTaxLiabilityService
 import uk.gov.hmrc.play.test.UnitSpec
 
 import scala.concurrent.Future
 
-sealed trait MockServices extends UnitSpec with MockitoSugar with BeforeAndAfterEach
 
-trait MockEstimatedLiabilityService extends MockServices {
+trait MockEstimatedLiabilityService extends UnitSpec with MockitoSugar with BeforeAndAfterEach {
 
   val mockEstimatedLiabilityService: EstimatedTaxLiabilityService = mock[EstimatedTaxLiabilityService]
 
@@ -40,20 +39,4 @@ trait MockEstimatedLiabilityService extends MockServices {
   def setupMockLastTaxCalculationResult(mtditid: String)(response: LastTaxCalculationResponseModel): Unit =
     when(mockEstimatedLiabilityService.getLastEstimatedTaxCalculation(ArgumentMatchers.eq(mtditid))(ArgumentMatchers.any()))
       .thenReturn(Future.successful(response))
-}
-
-
-trait MockObligationsService extends MockServices {
-
-  val mockObligationsService: ObligationsService = mock[ObligationsService]
-
-  override def beforeEach(): Unit = {
-    super.beforeEach()
-    reset(mockObligationsService)
-  }
-
-  def setupMockObligationsResult(nino: String)(response: ObligationsResponseModel): Unit = {
-    when(mockObligationsService.getObligations(ArgumentMatchers.eq(nino))(ArgumentMatchers.any()))
-      .thenReturn(Future.successful(response))
-  }
 }
